Fix the unit test of Storyboard Reference on OS X.
The previously implemented test did not actually run when I was using OS X 10.10. After I update my Mac to 10.11, the test started failing. The problem was deployment target version of the unit test for OS X, and update it to 10.11 (only in the test target). Setting a build version of a storyboard did not work for a compile error mentioned below.

http://stackoverflow.com/questions/31895295/ibtool-error-when-linking-storyboard-references-the-file-info-plist-couldnt