Enables support for distributing Swinject as a binary library
This fixes #450.

**Solution**

To allow frameworks to be used by multiple Swift version, we need to set:

```
BUILD_LIBRARY_FOR_DISTRIBUTION = YES
```

This isn't necessary for a Swift package (since they are build from source),
but it's useful for Carthage or other means of distributing binary frameworks.

This was added to the `project.yml` and then `xcodegen` was run.

**Testing**

I built `Swinject.framework` in Xcode 11.3.1. Then in Xcode 11.4 I created a
test project and linked to the generated project. It built and linked
successfully.